@@ -1,25 +1,28 @@
 module.exports = {
-  "parser": 'babel-eslint',
-  "env": {
-    "browser": true,
-    "commonjs": true,
-    "es6": true,
-    "node": true
+  parser: 'babel-eslint',
+  env: {
+    browser: true,
+    commonjs: true,
+    es6: true,
+    node: true,
+    jest: true
   },
-  "extends": ["eslint:recommended", "plugin:react/recommended"],
-  "parserOptions": {
-    "ecmaFeatures": {
-      "experimentalObjectRestSpread": true,
-      "jsx": true
+  extends: ['eslint:recommended', 'plugin:react/recommended'],
+  parserOptions: {
+    ecmaFeatures: {
+      jsx: true
     },
-    "sourceType": "module"
+    sourceType: 'module'
   },
-  "plugins": [ "react" ],
-  "rules": {
-    "indent": ["error", 2],
-    "linebreak-style": ["error","unix"],
-    "quotes": ["error","single"],
-    "semi": ["error","always"],
-    "no-console": ["warn", { "allow": ["info", "error"] }]
+  plugins: ['react'],
+  rules: {
+    indent: ['error', 2, { SwitchCase: 1 }],
+    'linebreak-style': ['error', 'unix'],
+    quotes: ['error', 'single'],
+    semi: ['error', 'always'],
+    'no-console': [
+      'warn',
+      { allow: ['clear', 'info', 'error', 'dir', 'trace'] }
+    ]
   }
 };
