@@ -1,15 +1,18 @@
-import React from 'react';
+import React, { Button } from 'react';
+import PropTypes from 'prop-types'
 
 const Header = ({ message }) => {
   return (
-    <h2 className="Header text-center">
-      {message}
-    </h2>
+    <div>
+      <h2 className="Header text-center">
+        {message}
+      </h2>
+    </div>
   );
 };
 
-Header.propTypes = {
-  message: React.PropTypes.string
-};
+// Header.propTypes = {
+//   message: React.propTypes.string
+// };
 
 export default Header;
