@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom';
 import App from './components/App';
 
 
-ReactDOM.render(
+ReactDOM.hydrate(
   <App initialData={window.initialData} />,
   document.getElementById('root')
 );
